requirements.txt
streamlit>=1.22.0
google-generativeai>=0.3.0
python-dotenv
PyPDF2>=3.0.0
python-docx>=0.8.11